Handle false in configurations files
Previously setting a configuration key to have a false value raised an
error that the hash doesn't respond to the given value, even though the
key and value are defined in the configuration. IE:

config = {:foo => false}
config.extend(MagicHash)
config.foo # raises NoMethodError (#foo not found on Hash)

This patch fixes this issue.