Fixed the overlaying of values in local config
The original code didn't OVERLAY the local config with the standard config
and there was no real mention of it in the docs. So I added a test, it failed,
and then I fixed the code. The simple thing was to merge the Hash as opposed
to replacing it. Now it works as you'd expect it.